optimise: restructurise project files - add data folder
