+++ components/FAQ/fag-list.tsx
@@ -1,5 +1,5 @@
 import { IQuestions } from 'types';
-import { getFAQQuestions } from 'helpers';
+import { getFAQQuestions } from 'data';
 
 import { Container } from 'layout';
 import { MappedListTemplate } from 'template';
