+++ components/home/feedback.tsx
@@ -1,6 +1,7 @@
 import { SectionTitle } from 'types';
+import { getFeedbackSlides } from 'data';
 
-import { feedbackBreakpoints, getFeedbackSlides, getIdValues } from 'helpers';
+import { feedbackBreakpoints, getIdValues } from 'helpers';
 
 import { SectionTemplate } from 'template';
 import { CarouselUI } from 'ui';
