+++ components/home/overview.tsx
@@ -1,6 +1,7 @@
 import { IOrderStep, SectionTitle } from 'types';
+import { getOrderSteps } from 'data';
 
-import { getOrderSteps, getIdValues, imageSettings } from 'helpers';
+import { getIdValues, imageSettings } from 'helpers';
 
 import { MappedListTemplate, SectionTemplate } from 'template';
 import { OverviewItem } from './subcomponents';
