+++ components/home/subcomponents/benefits-list.tsx
@@ -1,5 +1,5 @@
 import { IBenefitsItem } from 'types';
-import { getBenefits } from 'helpers';
+import { getBenefits } from 'data';
 
 import { MappedListTemplate } from 'template';
 import BenefitsItem from './benefits-item';
