+++ components/layout/footer-navigation.tsx
@@ -4,8 +4,9 @@ import Link from 'next/link';
 import { usePathname } from 'next/navigation';
 
 import { Paths, PositionInLayout } from 'types';
+import { getFooterLinks } from 'data';
 
-import { getFooterLinks, mapArray } from 'helpers';
+import { mapArray } from 'helpers';
 
 import { CalculationModalTrigger } from './subcomponents';
