+++ components/performers/about-us.tsx
@@ -1,6 +1,6 @@
 import { IAboutUs } from 'types';
 
-import { getAboutUsData } from 'helpers';
+import { getAboutUsData } from 'data';
 
 import { Container } from 'layout';
 import { MappedListTemplate } from 'template';
