+++ helpers/index.ts
@@ -1,14 +1,10 @@
 export * from './anchorLinks';
-export * from './calculationData';
 export * from './calculationHelper';
 export * from './carouselHelper';
 export * from './cookiesHelper';
-export * from './componentsData';
 export * from './imageSettings';
-export * from './layoutsData';
 export * from './layoutsHelper';
 export * from './mapArray';
-export * from './metadata';
 export * from './sectionHelper';
 export * from './tailwindPlugins';
 export * from './telegramHelpers';
