+++ hooks/useActiveLink.ts
@@ -5,7 +5,7 @@ import { usePathname } from 'next/navigation';
 import { debounce } from 'lodash';
 
 import { Paths } from '../types';
-import { getAdaptedLinks } from 'helpers';
+import { getAdaptedLinks } from '../data';
 import { useIntersectionObserver } from './useIntersectionObserver';
 
 export const useActiveLink = (isDesktop: boolean) => {
