+++ tsconfig.json
@@ -39,6 +39,7 @@
     "paths": {
       "components/*": ["components/*"],
       "context/*": ["context/*"],
+      "data/*": ["data/*"],
       "helpers/*": ["helpers/*"],
       "hooks/*": ["hooks/*"],
       "layout/*": ["layout/*"],
